Grouped models with relationship into one mode
