+++ agriTech/agriTech/settings.py
@@ -24,9 +24,6 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'lending_records',
-    'officer_visits',
-    'tracking',
     'machinery',
     'users',
     'api',
