--- agriTech/lending_records/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-from .models import Lending_Record
-admin.site.register(Lending_Record)
