--- agriTech/officer_visits/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-from .models import Officer_Visit
-admin.site.register(Officer_Visit)
