--- agriTech/officer_visits/views.py
@@ -1,4 +1,0 @@
-
-from django.shortcuts import render
-
-# Create your views here.
