--- agriTech/tracking/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-from .models import Machinery_Tracking
-admin.site.register(Machinery_Tracking)
