--- agriTech/tracking/urls.py
@@ -1,2 +1,0 @@
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
