--- agriTech/tracking/views.py
@@ -1,3 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
